requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
openai>=1.40.0
pymupdf>=1.24.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.1
python-docx>=0.8.11
